Refactored all routers to move controller functions to separate file
