+++ server/routes/users.js
@@ -1,22 +1,14 @@
 const Router = require("express-promise-router");
 const User = require("../models/User");
 const Activity = require("../models/Activity")
+const usersController = require("../controllers/users")
 const router = new Router();
 
-router.get("/:user_id/activities", async (req, res) => {
-  console.log("GET ALL USER ACTIVITIES")
-})
 
-router.get("/:user_id", async (req, res) => {
-  console.log("GET  USER ")
+router.get("/:user_id/activities", usersController.getActivities )
 
-})
+router.get("/:user_id", usersController.get)
 
-router.get("/", async (req, res) => {
-  console.log("GET ALL USERS")
-  // get all addresses
-  const users = await User.all()
-  res.status(200).json(users)
-});
+router.get("/", usersController.getAll);
 
 module.exports = router
